feat: add info to clarify vendor issues
Unfortunately, we often receive issue reports that we can't do anything about,
and should be sent to the EVS vendor instead. Add clarification text in
the `/feedback` handler where it is most likely to be seen, as well as an
additional command linking the project for those who are curious.

Note that the additional command will need to be registered separately on
Telegram via BotFather.