build: remove uptime-kuma
Remove uptime-kuma because we don't want the monitoring system to live on the
system that we are monitoring in the first place. Monitoring has been replaced
by alerts set up on DigitalOcean, where the bot is currently being hosted.

While we're updating the Compose file, remove the `version` field, which has
reportedly been deprecated.